Update Decoder to support repeated numeric types
From the docs: (https://developers.google.com/protocol-buffers/docs/encoding#packed)

Protocol buffer parsers must be able to parse repeated fields that
were compiled as packed as if they were not packed, and vice versa.
This permits adding [packed=true] to existing fields in a forward- and
backward-compatible way.

Currently the Decoder does not support decoding a binary where the field
has been packed, but the module defines a field which has not been
tagged with the [packed=true] option. In that case the "wrong wire_type
for #{field}" error is shown. This behaviour does not conform with the
recomendation from the docs (as shown above).

Therefore this commit updates the Decoder to support both packed and
unpacked binaries for proto2, using a module where the packed=true
option has been set, and one where it has not been set.

Note: For proto3 this is a non issue, because the default behaviour is packed.