Fix Decoder UndefinedFunctionError
Previously when creating the error message as a result of decoding
a delimited field and running out of data, the message attempted to
include the field name, however sometimes this value is not available in
the props and instead results in an unhelpful undefined function error.
This commit updates the code to ensure the attempt to locate the name
fails gracefully and simply provides the field number when name is not
available.